remove 64KB limitation for payload

+++ api/oc_collection.c
@@ -993,7 +993,7 @@ oc_handle_collection_request(oc_method_t method, oc_request_t *request,
     }
   }
   request->response->response_buffer->content_format = APPLICATION_VND_OCF_CBOR;
-  request->response->response_buffer->response_length = (uint16_t)size;
+  request->response->response_buffer->response_length = size;
   request->response->response_buffer->code = code;
 
   if ((method == OC_PUT || method == OC_POST) &&
